refactor: move ui events to settings.js
